merge

+++ course/models.py
@@ -8,7 +8,7 @@ class Course(models.Model):
     title = models.CharField(max_length=200)
 
     description = models.TextField()
-    course_fee = models.DecimalField(decimal_places=2, max_digits=10, default=0)
+    course_fee = models.IntegerField( default=0)
     course_length = models.CharField(max_length=200, default='')
     video_thumbnail_url = models.CharField(max_length=200, default='')
     video_playlist_url = models.CharField(max_length=200, default='')
